Do not store WorkerExecutorFacade as a field
Instantiate WorkerExecutorFacade only when necessary and do not store
it in a field. This is added to AndroidVariantTask so all tasks have
access to worker executor. It handles injections of WorkerExecutor, and
it provides default method for getting the WorkerExecutorFacade.

This change is also necessary to allow progress with instant execution
as serialization of WorkerExecutorFacade field fails.

BUG: 137195497
Test: TaskMethodModifiersAndAnnotationsTest
Change-Id: If63f5487ff078387e35bbf20b6086b4dde0f87e4