Use application class loader for hot swap
If the android:sharedUserId or android:process attributes are set on an
application, the thread context class loader will be replaced with an
empty dummy classloader.

In addition, it is possible for users to modify the thread context
classloader. However, the application class loader should contain all
classes present in the app.

Bug: 135172147
Bug: 117167542
Test: Existing coverage + manual
Change-Id: Ic642fce78a42d2f1224ba866d2f2c420662f676c