Fix input normalization of AidlCompile task.
Given the same version, the path or contents of the AIDL tool may change
across platforms, but it would still produce the same output (given the
same inputs). Therefore, the path of contents of the tool should not be
an input. Instead, we set the tool's version as input.

(The above probably applies to all build tools in general and not just
AIDL.)

Fixes: 138920846
Test: Existing tests should pass
Change-Id: Iefcdfb2be0ad5051a71c9ea29cb8f83d5adcb150