Remove unnecessary Task.dependsOn for connected android test
When configuring DeviceProviderInstrumentTestTask with Gradle 5.6 there
is an issue https://github.com/gradle/gradle/issues/10177. Task dependencies
cannot be computed because the Task.dependsOn is invoked with an empty
provider when configuring this task for a library project.

However, this invocation of Task.dependsOn is redundant as tested APKs
are declared as input of this task, so this CL removes that invocation.

BUG: Gradle issue in the description
Test: BuilderTestingApiTest and Gradle 5.6
Change-Id: Ie4530b40f506fdb9b84773ccb9d6eb2ae7b70c3a