Fix SymbolTableBuilder sometimes dropping symbols.
It would previously drop the last symbol if it was a styleable symbol,
as the styleable is not added to the symbol table until the next symbol
is added.

This happens to be the case in androidx.cardview:cardview:1.0.0,
see https://issuetracker.google.com/139017699#comment9 for details.

The fix is to do this check both on next sybmol and at the end.

Bug: 139017699
Test: Added SymbolUtilsTest.testSymbolTableBuilderStyleableLast()
Change-Id: I0a56e05faa83506a02775f9f447bef0071bfce83