Return proper error on agent setup failure
Updates agent.cc to return a proper response if any part of setup other
than the socket connection fails.

Refactors swapper.cc and agent.cc to achieve this in a cleaner fashion;
agent.cc is now responsible for sending responses. This additionally
reduces the number of unique_ptr that need to be passed around.

Fixes: 139815748
Test: Manual
Change-Id: I21736840141d1aa8f3a96ff56cb4389f07e4cd20