Escape expected string in ProcessRunner.waitForInput
In our test we print an array to System.out and
waitForInput in the test. So expected string was [1, 2, 127]
test didn't successfully wait for it because "[]" are special
symbols in regexp.

As a drive-by:
1. use long instead of int for timeout in millis, because all java api
related to millis always use long, for example TimeUnit.SECONDS.toMillis
return long value

2. renamed timeout parameter to timeoutMs so callers don't have to
guess and verify that they should pass value in milliseconds

3. AppInspectionTest: fix waitForInput usage asserting that input was
received

4. AppInspectionTest: fix sent commandId

Test: bazel test //tools/base/profiler/tests/perf-test:all
bug: 137542165
Change-Id: I371268fb6fefe7f48faf59666f2811f4f2600ec4