Avoid writing signing config info to disk when possible.
To protect sensitive data, a packaging task in the AGP should get the
signing config information
  - directly from the variant scope of a non-dynamic-feature module, or
  - directly from the injected properties passed by the IDE

If the above are not possible, the dynamic-feature module will get this
information from a file published by the base module as it used to do.
We'll consider encrypting this file in a separate CL.

Note: As a side effect, this CL will solve a deletion issue on the
signing config file for projects that are not using dynamic features
(see bug 137448928).

Bug: 137210434
Bug: 137448928
Test: New tests in SigningTest, DynamicAppSigningConfigTest
Change-Id: Iead623b90b290938168835a9d442b38b96672de9