Add more fields to the test layout
After updating aapt2, the compiled layout size was smaller than the
replacement that shrink resources task was replacing it with, so the
task was not stripping out the layout. This change adds more fields to
the layout so the compiled size is more than the replacement size so
that shrink resources task strips it out.

Test: tests are passing
Bug: N/A
Change-Id: I2ea90926c484cf1dc2315c1d7b16f257c23b7975