add transformAll implementation to VariantAPI.
transform on multiple artifacts always consume a list of
existing producers and produce a single output which
will be turned into a list of a single element for the next
transform (if any).

if at least one transform is registered, the final artifact
will always be a list of a single element (the last transform
output).

rename AppendRequest.to to AppendRequest.on to avoid colliding
with Kotlin infix to notation.

Test: added unit tests
Bug: 139752844

Change-Id: Ic4190b077beddebf6ae38512aa05b5ac43b92e74