Don't remove native libs when external libs are updated
Previously, when updating an external lib with native libs, sometimes
some of the native libs would be removed from the APK. The reason for
this is because the IncrementalFileMerger assumed that the absolute
paths of input directories wouldn't change between builds, but the
absolute paths of the input directories do change when the version
changes.

This CL changes the IncrementalFileMerger logic such that it no longer
relies on the absolute paths of the input directories being constant.

This CL is a revision of I2f07b868b25d19527d25a551e5cb3a9fe5cc82d9,
which caused a regression. The regression was caused because
LazyIncrementalFileMergerInput.getAllPaths() is an expensive call, and
that CL was calling it for all inputs. This CL avoids calling it in the
vast majority of cases.

Fixes: 141536986
Test: IncrementalFileMergerTests, NativeSoPackagingFromRemoteAarTest
Change-Id: I3c641279764c2fd054d0ec5f2054561e04caca8d